+++ src/components/MyEditor/index.js
@@ -2,9 +2,10 @@ import React, { useRef } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import styled from "styled-components";
 
+import { selectors, actions } from "data";
+
 import BasicEditor from "./BasicEditor";
 import Controller from "./Controller";
-import { selectors, actions } from "data";
 import { saveContent } from "./helper";
 
 export default function MyEditor({
@@ -14,54 +15,21 @@ export default function MyEditor({
   id
 }) {
   const dispatch = useDispatch();
-  const postUserId = useSelector(selectors.post.getCurrentPostUserId);
-  const isSameUser = id ? id === postUserId : false;
-
   const editorRef = useRef();
+  const userId = useSelector(selectors.user.getUserId);
 
-  const saveHandler = () => {
-    saveContent({ editorState, id });
+  const saveHandler = (editorState, userId) => {
+    saveContent({ editorState, id: userId });
     dispatch(actions.modal.modalUpAndGo("saved!"));
-  };
-  const publishHandler = () => {
-    dispatch(actions.post.createPost());
-    dispatch(actions.modal.modalUpAndGo("published!"));
-  };
-
-  const editHandler = postId => {
-    dispatch(actions.router.push(`/postEdit/${postId}`));
-  };
-
-  const confirmEdithandler = () => {
-    dispatch(actions.post.updatePost());
-  };
-
-  const deleteHandler = postId => {
-    dispatch(
-      actions.modal.setModalUp({
-        modalType: "DELETE_POST",
-        modalProps: { postId }
-      })
-    );
-  };
-
-  const handlers = {
-    saveHandler,
-    publishHandler,
-    editHandler,
-    deleteHandler,
-    confirmEdithandler
   };
 
   return (
     <>
-      <Controller handlers={handlers} isSameUser={isSameUser} />
-
       <BasicEditor
         editorState={editorState}
         setEditorState={setEditorState}
         editorRef={editorRef}
-        saveHandler={saveHandler}
+        saveHandler={saveHandler.bind(this, editorState, userId)}
         readOnly={readOnly}
       />
     </>
